Move gsCDRAssembler::assemble to hpp file
