+++ src/gsAssembler/gsCDRAssembler_.cpp
@@ -1,0 +1,10 @@
+
+#include <gsCore/gsTemplateTools.h>
+
+#include <gsAssembler/gsCDRAssembler.h>
+#include <gsAssembler/gsCDRAssembler.hpp>
+
+namespace gismo
+{
+    CLASS_TEMPLATE_INST gsCDRAssembler<real_t>;
+}
